refactor(utils): move sourcemap generation into addNewLine function
